Importando cuotas reales

+++ MyBettingAI/Models/Team.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string Name { get; set; } = "";
         public int LeagueId { get; set; }
-        public double StrengthRating { get; set; }
+        public float StrengthRating { get; set; }
         public int? ApiId { get; set; }
     }
 }
